Futher cleanup
